no truncate, 0 for success, add err checking to adapt

+++ src/utils/c_utils/inc/aduc/string_c_utils.h
@@ -34,7 +34,7 @@ bool IsNullOrEmpty(const char* str);
 
 bool MallocAndSubstr(char** target, char* source, size_t len);
 
-size_t ADUC_Safe_StrCopyN(char* dest, const char* src, size_t destByteLen, size_t numSrcCharsToCopy);
+int ADUC_Safe_StrCopyN(char* dest, const char* src, size_t destByteLen, size_t numSrcCharsToCopy);
 
 int ADUC_AllocAndStrCopyN(char** dest, const char* src, size_t srcLen);
